[html layout] Suggest tutorial readers to download code of moveit_tutorials (address the concern raised in https://github.com/ros-planning/moveit_tutorials/pull/64).
We noticed that some tutorial resources are duplicated, in the form of source and binary. We will want to only use source and https://github.com/ros-planning/moveit_tutorials as a single location to maintain all tutorial resource.

This PR tries to show "note" section at the top of every tutorial page to suggest users to clone the source.

Hope this is close to the solution what we're discussing in #64.