Removed references to the panda arm that only made sense for the pr2. Some typos.
In some places it does not make sense to replace the pr2 with the panda arm because the panda arm is fixed to the environment (see line #57) and does not have a left and a right arm (see line #81). I would suggest replacing panda with a mobile robot (e.g. the pr2) in line #57 and using less specific termes like left_arm and right_arm in line #81. Also, there were two typos that ocurred to me when reading the tutorial.