Fixed download of rosinstall
Previously the clone of moveit-master, then merge of rosinstall would modify the branch during the merge and update. It makes more sense to point to the correct file of git for the rosinstall download. It is also more flexible for the future stable branch if such a tutorial will be made for it.