Update realtime_servo_tutorial.rst
Although, the code suggests to clone the old version of UR's repo but it does not clearly mention it in words. I did not clone the old version because I already had the latest version and the tutorial didn't work for me for now obvious reasons. So, I am adding a note in bold so that no one else does the same mistake as mine and don't waste 2 days like me. I also added a debug tip to see when one should expect the tutorial to work.